change the repository structure to contain multiple packages
