+++ scripts/format
@@ -1,4 +1,6 @@
 #!/usr/bin/env bash
 
-find src -type f -name "*.hs" | xargs ormolu -i
-find test -type f -name "*.hs" | xargs ormolu -i
+REPO_DIR=$( cd -- "$( dirname -- "${BASH_SOURCE[0]}" )/.." &> /dev/null && pwd )
+
+find $REPO_DIR/secp256k1-haskell/src -type f -name "*.hs" | xargs ormolu -i
+find $REPO_DIR/secp256k1-haskell/test -type f -name "*.hs" | xargs ormolu -i
