+++ secp256k1-haskell/LICENSE
@@ -1,0 +1,1 @@
+../LICENSE
