+++ secp256k1-haskell/README.md
@@ -1,0 +1,1 @@
+../README.md
