+++ secp256k1-haskell/package.yaml
@@ -1,5 +1,5 @@
 name: secp256k1-haskell
-version: 1.0.0
+version: 1.0.1
 synopsis: Bindings for secp256k1
 description: Sign and verify signatures using the secp256k1 library.
 category: Crypto
@@ -22,7 +22,7 @@ dependencies:
   - hashable >=1.2.6 && <1.5
   - QuickCheck >=2.9.2 && <2.15
   - string-conversions >=0.4 && <0.5
-  - unliftio-core >=0.1.0 && <0.3 
+  - unliftio-core >=0.1.0 && <0.3
 library:
   source-dirs: src
   generated-exposed-modules:
