+++ secp256k1-haskell/src/Crypto/Secp256k1/Internal/BaseOps.hs
@@ -1,7 +1,5 @@
-{-# LANGUAGE ImportQualifiedPost #-}
-
 -- |
--- Module      : Crypto.Secp256k1.Internal
+-- Module      : Crypto.Secp256k1.Internal.BaseOps
 -- License     : UNLICENSE
 -- Maintainer  : Jean-Pierre Rupp <jprupp@protonmail.ch>
 -- Stability   : experimental
@@ -9,128 +7,28 @@
 --
 -- The API for this module may change at any time. This is an internal module only
 -- exposed for hacking and experimentation.
-module Crypto.Secp256k1.Internal where
+module Crypto.Secp256k1.Internal.BaseOps where
 
-import Data.ByteString (ByteString)
-import Data.ByteString qualified as BS
-import Data.ByteString.Unsafe qualified as BU
-import Foreign (FunPtr, Ptr, castPtr)
+import Crypto.Secp256k1.Internal.ForeignTypes
+  ( Compact64,
+    LCtx,
+    Msg32,
+    NonceFun,
+    PubKey64,
+    Ret,
+    SecKey32,
+    Seed32,
+    SerFlags,
+    Sig64,
+    Tweak32,
+  )
+import Foreign (FunPtr, Ptr)
 import Foreign.C
   ( CInt (..),
     CSize (..),
-    CString,
     CUChar,
     CUInt (..),
   )
-import GHC.Generics (Selector (selDecidedStrictness))
-import System.Entropy (getEntropy)
-
-data LCtx
-
-data PubKey64
-
-data Msg32
-
-data Sig64
-
-data Seed32
-
-data Compact64
-
-data SecKey32
-
-data Tweak32
-
-type CtxFlags = CUInt
-
-type SerFlags = CUInt
-
-type Ret = CInt
-
-type NonceFun a =
-  Ptr CUChar ->
-  Ptr CUChar ->
-  Ptr CUChar ->
-  Ptr CUChar ->
-  Ptr a ->
-  CInt ->
-  IO CInt
-
-verify :: CtxFlags
-verify = 0x0101
-
-sign :: CtxFlags
-sign = 0x0201
-
-signVerify :: CtxFlags
-signVerify = 0x0301
-
-compressed :: SerFlags
-compressed = 0x0102
-
-uncompressed :: SerFlags
-uncompressed = 0x0002
-
-isSuccess :: Ret -> Bool
-isSuccess 0 = False
-isSuccess 1 = True
-isSuccess n = error $ "isSuccess expected 0 or 1 but got " ++ show n
-
-unsafeUseByteString :: ByteString -> ((Ptr a, CSize) -> IO b) -> IO b
-unsafeUseByteString bs f =
-  BU.unsafeUseAsCStringLen bs $ \(b, l) ->
-    f (castPtr b, fromIntegral l)
-
-useByteString :: ByteString -> ((Ptr a, CSize) -> IO b) -> IO b
-useByteString bs f =
-  BS.useAsCStringLen bs $ \(b, l) ->
-    f (castPtr b, fromIntegral l)
-
-unsafePackByteString :: (Ptr a, CSize) -> IO ByteString
-unsafePackByteString (b, l) =
-  BU.unsafePackMallocCStringLen (castPtr b, fromIntegral l)
-
-packByteString :: (Ptr a, CSize) -> IO ByteString
-packByteString (b, l) =
-  BS.packCStringLen (castPtr b, fromIntegral l)
-
-withRandomSeed :: (Ptr Seed32 -> IO a) -> IO a
-withRandomSeed go = do
-  bs <- getEntropy 32
-  useByteString bs $ go . fst
-
-foreign import ccall safe "secp256k1.h secp256k1_context_create"
-  contextCreate ::
-    CtxFlags ->
-    IO (Ptr LCtx)
-
-foreign import ccall safe "secp256k1.h secp256k1_context_clone"
-  contextClone ::
-    Ptr LCtx ->
-    IO (Ptr LCtx)
-
-foreign import ccall safe "secp256k1.h secp256k1_context_destroy"
-  contextDestroy ::
-    Ptr LCtx ->
-    IO ()
-
-foreign import ccall safe "secp256k1.h secp256k1_context_set_illegal_callback"
-  setIllegalCallback ::
-    Ptr LCtx ->
-    -- | message, data
-    FunPtr (CString -> Ptr a -> IO ()) ->
-    -- | data
-    Ptr a ->
-    IO ()
-
-foreign import ccall safe "secp256k1.h secp256k1_context_set_error_callback"
-  setErrorCallback ::
-    Ptr LCtx ->
-    -- | message, data
-    FunPtr (CString -> Ptr a -> IO ()) ->
-    -- | data
-    Ptr a ->
-    IO ()
 
 foreign import ccall safe "secp256k1.h secp256k1_ec_pubkey_parse"
   ecPubKeyParse ::
@@ -260,12 +158,6 @@ foreign import ccall safe "secp256k1.h secp256k1_ec_pubkey_tweak_mul"
     Ptr LCtx ->
     Ptr PubKey64 ->
     Ptr Tweak32 ->
-    IO Ret
-
-foreign import ccall safe "secp256k1.h secp256k1_context_randomize"
-  contextRandomize ::
-    Ptr LCtx ->
-    Ptr Seed32 ->
     IO Ret
 
 foreign import ccall safe "secp256k1.h secp256k1_ec_pubkey_combine"
