+++ stack.yaml
@@ -1,7 +1,12 @@
 resolver: lts-21.4
+
+packages:
+- secp256k1-haskell
+
 nix:
   packages:
     - secp256k1
     - pkg-config
+
 extra-deps:
   - base16-1.0@sha256:9b72a280a7af75a5026fa25a1b8ae18ec10200a070947723f1fd61dc8d407862,2472
